兼容 oracle 改造

+++ api/src/main/java/top/cardone/configuration/po/I18nInfo.java
@@ -141,7 +141,7 @@ public class I18nInfo implements java.io.Serializable {
      */
     @lombok.Getter
     @lombok.Setter
-    protected Long orderBy;
+    protected Integer orderBy;
 
     /**
      * 组织编号
