+++ api/src/main/java/top/cardone/configuration/po/SystemInfo.java
@@ -127,7 +127,7 @@ public class SystemInfo implements java.io.Serializable {
      */
     @lombok.Getter
     @lombok.Setter
-    protected Long orderBy;
+    protected Integer orderBy;
 
     /**
      * 组织编号
